requirements.txt
Django==3.0.2; python_version > '3.5'
Django==2.2.9; python_version == '3.5'
Django==1.11.28;python_version < '3.0'
# Uncomment the mysqlclient requirement if you are using MySQL rather than
# PostgreSQL. You must also have a MySQL client installed in that case.
mysqlclient==1.4.1
wheel==0.36.2
gunicorn==20.0.4; python_version > '3.0'
gunicorn==19.10.0; python_version < '3.0'
psycopg2==2.8.4
djangorestframework
